moved the available rooms up a lvl
